Conflicts:
	packages/zoltan2/src/algorithms/order/Zoltan2_AlgAMD.hpp
	packages/zoltan2/src/environment/Zoltan2_InputTraits.hpp
	packages/zoltan2/src/input/Zoltan2_MeshAdapter.hpp
	packages/zoltan2/src/input/Zoltan2_PamgenMeshAdapter.hpp
	packages/zoltan2/src/models/Zoltan2_CoordinateModel.hpp
	packages/zoltan2/src/problems/Zoltan2_PartitioningProblem.hpp
	packages/zoltan2/src/problems/Zoltan2_PartitioningSolution.hpp
	packages/zoltan2/test/unit/problems/PartitioningSolution.cpp

+++ packages/seacas/libraries/aprepro_lib/apr_builtin.cc
@@ -1127,7 +1127,7 @@ array *do_csv_array1(const char *filename)
 
 array *do_csv_array(const char *filename, double skip)
 {
-  int rows_to_skip = (int)skip;
+  size_t rows_to_skip = (size_t)skip;
   
   const char *delim = ",\t ";
   std::fstream *file = aprepro->open_file(filename, "r");
@@ -1167,7 +1167,7 @@ array *do_csv_array(const char *filename, double skip)
       }
     }
   }
-  assert((int)rows - rows_to_skip == array_data->rows);
+  assert(rows - rows_to_skip == (size_t)array_data->rows);
   return array_data;
 }
